+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "keksobooking",
   "version": "26.0.0",
-  "private": true,
+  "type":"module",
   "description": "Личный проект «Кексобукинг» от HTML Academy",
   "repository": {
     "type": "git",
